+++ examples/api500/storage_system.rb
@@ -21,13 +21,18 @@ require_relative '../_client' # Gives access to @client
 # - API200 for C7000 and Synergy (see /examples/shared_samples/storage_system.rb)
 # - API300 for C7000 and Synergy (see /examples/shared_samples/storage_system.rb)
 # - API500 for C7000 and Synergy
+# - API600 for C7000 and Synergy (see /examples/api600/storage_system.rb)
 
 raise 'ERROR: Must set @storage_system_ip in _client.rb' unless @storage_system_ip
 raise 'ERROR: Must set @storage_system_username in _client.rb' unless @storage_system_username
 raise 'ERROR: Must set @storage_system_password in _client.rb' unless @storage_system_password
 
 if @client.api_version < 500
-  raise "If you want execute sample for API < #{@client.api_version}, you should execute the ruby file '/examples/shared_samples/storage_system.rb'"
+  raise "If you want execute sample for API < #{@client.api_version}," \
+         "you should execute the ruby file '/examples/shared_samples/storage_system.rb'"
+elsif @client.api_version == 600
+  raise "If you want execute sample for API #{@client.api_version}," \
+         "you should execute the ruby file '/examples/api600/storage_system.rb'"
 end
 
 # Resources classes that you can use for Storage System in this example:
@@ -36,27 +41,27 @@ end
 storage_system_class = OneviewSDK.resource_named('StorageSystem', @client.api_version)
 
 # for StorageSystem with family StoreServ
-# options = {
-#   credentials: {
-#     username: @storage_system_username,
-#     password: @storage_system_password
-#   },
-#   hostname: @storage_system_ip,
-#   family: 'StoreServ',
-#   deviceSpecificAttributes: {
-#     managedDomain: 'TestDomain'
-#   }
-# }
-
-# for StorageSystem with family StoreVirtual
 options = {
   credentials: {
     username: @storage_system_username,
     password: @storage_system_password
   },
   hostname: @storage_system_ip,
-  family: 'StoreVirtual'
+  family: 'StoreServ',
+  deviceSpecificAttributes: {
+    managedDomain: 'TestDomain'
+  }
 }
+
+# for StorageSystem with family StoreVirtual
+# options = {
+#   credentials: {
+#     username: @storage_system_username,
+#     password: @storage_system_password
+#   },
+#   hostname: @storage_system_ip,
+#   family: 'StoreVirtual'
+# }
 
 storage_system = storage_system_class.new(@client, options)
 puts "\nAdding a storage system with"
@@ -73,7 +78,6 @@ end
 
 storage_system = storage_system_class.new(@client, hostname: storage_system['hostname'])
 storage_system.retrieve!
-
 port = storage_system['ports'].find { |item| item['protocolType'].downcase.include?('fc') } # find first correct protocolType for using our fc network
 if port
   fc_network = OneviewSDK::API500::C7000::FCNetwork.get_all(@client).first